requirements.txt
pydantic
pydantic_settings
fastapi
jinja2
uvicorn[standard]
httpx
# modules
python-jose[cryptography]
aiohttp
# testing
pytest
pytest-cov
pytest-asyncio